+++ src/main/kotlin/ru/yandex/money/tools/grafana/dsl/variables/Variable.kt
@@ -16,5 +16,8 @@ interface Variable : Json<JSONObject> {
      */
     val name: String
 
-    fun asVariable(): String
+    /**
+     * Returns string representation of this variable.
+     */
+    fun asVariable(): String = "$$name"
 }